requirements.txt
# csvwise dependencies
streamlit>=1.30.0
pandas>=2.0.0
matplotlib>=3.7.0
psycopg2-binary>=2.9.0  # PostgreSQL support (optional)
